UPDATE: /auth dir was renamed to /users FIX: import paths UPDATE: get_current_active_user, get_current_active_admin deps moved to auth_deps.py in global deps dir UPDATE: ORM UserModel moved from global /database/models.py to /users/models.py UPDATE: 'Roles' enum class moved from /database/models.py to separate /database/enums.py file FIX: paths in alembic 'env.py' file
