+++ api/src/database/enums.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class Roles(str, enum.Enum):
+    SUPER_ADMIN = "SUPER_ADMIN"
+    ADMIN = "ADMIN"
+    USER = "USER"
