+++ api/src/database/models.py
@@ -1,9 +1,8 @@
 import datetime
 import uuid
-import enum
 from typing import Annotated
 
-from sqlalchemy import VARCHAR, DateTime, BOOLEAN, Enum, text
+from sqlalchemy import DateTime, text
 from sqlalchemy.orm import DeclarativeBase, mapped_column, Mapped
 from sqlalchemy.dialects.postgresql import UUID
 
@@ -19,12 +18,6 @@ uuid_pk = Annotated[
 int_pk = Annotated[int, mapped_column(primary_key=True, autoincrement=True, nullable=False)]
 
 
-class Roles(str, enum.Enum):
-    SUPER_ADMIN = "SUPER_ADMIN"
-    ADMIN = "ADMIN"
-    USER = "USER"
-
-
 class Base(DeclarativeBase):
     created_at: Mapped[datetime.datetime] = mapped_column(
         DateTime(timezone=True),
@@ -34,24 +27,4 @@ class Base(DeclarativeBase):
         DateTime(timezone=True),
         server_default=text("TIMEZONE('utc', now())"),
         onupdate=datetime.datetime.utcnow(),
-    )
-
-
-class UserModel(Base):
-    __tablename__ = "users"
-
-    id: Mapped[uuid_pk]
-    username: Mapped[str] = mapped_column(VARCHAR(32), unique=True, nullable=False)
-    email: Mapped[str] = mapped_column(VARCHAR(64), unique=True, nullable=False)
-    password: Mapped[str] = mapped_column(VARCHAR(64), nullable=False)
-    is_active: Mapped[bool] = mapped_column(BOOLEAN, server_default=text("TRUE"), nullable=False)
-    is_email_verified: Mapped[bool] = mapped_column(BOOLEAN, server_default=text("FALSE"), nullable=False)
-    role: Mapped[Roles] = mapped_column(
-        Enum(
-            Roles,
-            name="roles",
-            # native_enum=False указывает не создавать Enum на уровне СУБД (хранится как varchar)
-        ),
-        server_default=text("'USER'::roles"),
-        nullable=False,
     )
