+++ api/src/users/dependencies.py
@@ -8,14 +8,13 @@ from jose import JWTError
 from redis.asyncio import Redis
 
 from .repository import UserRepository
-from .service import AuthService, UserService
+from .services import AuthService, UserService
 from .utils import decode_jwt, validate_token_type
 from .schemas import TokenData, BaseUserInfo
-from .exceptions import HTTPExceptionInvalidToken, HTTPExceptionInactiveUser, HTTPExceptionNoPermission
+from .exceptions import HTTPExceptionInvalidToken
 
 from api.src.config import settings
-from api.src.dependencies.db_dep import get_async_session_with_commit, get_async_redis_client
-from api.src.database.models import Roles
+from api.src.dependencies.db_deps import get_async_session_with_commit, get_async_redis_client
 
 logger = logging.getLogger("my_app")
 
@@ -88,23 +87,3 @@ def get_auth_dependency_from_token_type(token_type: str) -> callable:
 
 get_current_auth_user_by_access = get_auth_dependency_from_token_type(settings.ACCESS_TOKEN_NAME)
 get_current_auth_user_by_refresh = get_auth_dependency_from_token_type(settings.REFRESH_TOKEN_NAME)
-
-
-async def get_current_active_user(
-        user: Annotated[BaseUserInfo, Depends(get_current_auth_user_by_access)],
-) -> BaseUserInfo:
-    if not user.is_active:
-        logger.info(
-            f"Authorization: "
-            f"Inactive user {user.id} is trying to get in by access token!"
-        )
-        raise HTTPExceptionInactiveUser
-    return user
-
-
-async def get_current_active_admin(
-        user: Annotated[BaseUserInfo, Depends(get_current_active_user)],
-) -> BaseUserInfo:
-    if user.role not in [Roles.ADMIN, Roles.SUPER_ADMIN]:
-        raise HTTPExceptionNoPermission
-    return user
