+++ api/src/users/repository.py
@@ -1,5 +1,5 @@
 from api.src.database.repository import SQLAlchemyRepository
-from api.src.database.models import UserModel
+from api.src.users.models import UserModel
 
 
 class UserRepository(SQLAlchemyRepository[UserModel]):
